#15 Adjust input parameters of create task API
