+++ src/main/java/com/webank/servicemanagement/dto/CreateTaskRequestInputDto.java
@@ -1,11 +1,10 @@
 package com.webank.servicemanagement.dto;
 
-public class CreateTaskRequest {
+public class CreateTaskRequestInputDto {
     private String callbackUrl;
     private String taskName;
     private String roleName;
     private String operator;
-    private String requestId;
 
     public String getCallbackUrl() {
         return callbackUrl;
@@ -37,13 +36,5 @@ public class CreateTaskRequest {
 
     public void setRoleName(String roleName) {
         this.roleName = roleName;
-    }
-
-    public String getRequestId() {
-        return requestId;
-    }
-
-    public void setRequestId(String requestId) {
-        this.requestId = requestId;
     }
 }
